+++ src/cmp/Communication.h
@@ -18,7 +18,7 @@ public:
 	~Communication() = default;
 
 	virtual Core::CoreStatus Initialize( void ) final override;
-	virtual void Update( const uint32_t currentTime ) final override;
+	virtual void Update( const uint64_t currentTime ) final override;
 
 	virtual Core::CoreStatus Send( Clusters::Frame &message ) final override;
