+++ src/cmp/ProximityInterface.h
@@ -9,7 +9,7 @@ public:
 	~ProximityInterface()      = default;
 
 	virtual Core::CoreStatus Initialize( void )       = 0;
-	virtual void Update( const uint32_t currentTime ) = 0;
+	virtual void Update( const uint64_t currentTime ) = 0;
 
 	virtual uint16_t GetDistance( void )  = 0;
 	virtual uint16_t GetThreshold( void ) = 0;
