+++ src/drv/DriverInterface.h
@@ -8,6 +8,6 @@ public:
 	~DriverInterface() = default;
 
 	virtual Core::CoreStatus Initialize( void )       = 0;
-	virtual void Update( const uint32_t currentTime ) = 0;
+	virtual void Update( const uint64_t currentTime ) = 0;
 };
 }
