+++ unittests/mock/cmp/MockSensorProximity.h
@@ -5,12 +5,12 @@
 namespace Component {
 class MockSensorProximity : public Component::SensorProximityInterface {
 public:
-	MOCK_METHOD0(Initialize, Core::CoreStatus(void) );
-	MOCK_METHOD1(Update, void( const uint32_t ) );
+	MOCK_METHOD0( Initialize, Core::CoreStatus( void ) );
+	MOCK_METHOD1( Update, void( const uint64_t ) );
 
-	MOCK_METHOD1(GetDistance, uint16_t(const SensorsId) );
-	MOCK_METHOD1(GetThreshold, uint16_t(const SensorsId) );
-	MOCK_METHOD2(SetThreshold, Core::CoreStatus(const SensorsId, uint16_t) );
-	MOCK_METHOD1(IsDetecting, bool(const SensorsId) );
+	MOCK_METHOD1( GetDistance, uint16_t( const SensorsId ) );
+	MOCK_METHOD1( GetThreshold, uint16_t( const SensorsId ) );
+	MOCK_METHOD2( SetThreshold, Core::CoreStatus( const SensorsId, uint16_t ) );
+	MOCK_METHOD1( IsDetecting, bool(const SensorsId) );
 };
 }
