+++ unittests/mock/drv/MockAdc.h
@@ -5,10 +5,10 @@
 namespace Driver {
 class MockAdc : public Driver::AdcInterface {
 public:
-	MOCK_METHOD0(Initialize, Core::CoreStatus( void ) );
-	MOCK_METHOD1(Update, void( const uint32_t ) );
-	
-	MOCK_METHOD0(StartConversion, void( void ) );
-	MOCK_METHOD0(Read, uint16_t(void) );
+	MOCK_METHOD0( Initialize, Core::CoreStatus( void ) );
+	MOCK_METHOD1( Update, void( const uint64_t ) );
+
+	MOCK_METHOD0( StartConversion, void( void ) );
+	MOCK_METHOD0( Read, uint16_t( void ) );
 };
 }
