+++ unittests/mock/drv/MockTick.h
@@ -6,9 +6,9 @@ namespace Driver {
 class MockTick : public Driver::TickInterface {
 public:
 
-	MOCK_METHOD0(GetUs, uint64_t(void) );
-	MOCK_METHOD0(GetMs, uint32_t(void) );
-	MOCK_METHOD1(DelayMs, void( uint16_t ) );
-	MOCK_METHOD1(DelayUs, void( uint32_t ) );
+	MOCK_METHOD0( GetUs, uint64_t( void ) );
+	MOCK_METHOD0( GetMs, uint64_t( void ) );
+	MOCK_METHOD1( DelayMs, void( uint64_t ) );
+	MOCK_METHOD1( DelayUs, void( uint64_t ) );
 };
 }
